refactor: Deduplicate fetch-actor-details handler logic
Extract shared handler body from default and internal variants into
buildFetchActorDetailsResult in fetch_actor_details_common.ts. Apply
dedent to multiline strings. Remove duplicated args schema from
internal tool.
